requirements.txt
gspread==5.12.4
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
openai==1.12.0
python-dotenv==1.0.1
nixpacks==1.0.0 